feat(queries): expose helper functions
These are useful for creating additional queries outside the library
