+++ src/index.js
@@ -1,9 +1,6 @@
 import {getQueriesForElement} from './get-queries-for-element'
 import * as queries from './queries'
-
-// exporting on the queries namespace as a convenience
-// in addition to exporting the queries themselves
-export {queries}
+import * as queryHelpers from './query-helpers'
 
 export * from './queries'
 export * from './wait'
@@ -12,6 +9,7 @@ export * from './matches'
 export * from './get-node-text'
 export * from './events'
 export * from './get-queries-for-element'
+export * from './query-helpers'
 export * from './pretty-dom'
 
 export {
@@ -19,4 +17,7 @@ export {
   // The new name is better. Remove this in the next major version bump.
   getQueriesForElement as bindElementToQueries,
   getQueriesForElement as within,
+  // export query utils under a namespace for convenience:
+  queries,
+  queryHelpers,
 }
