+++ typings/index.d.ts
@@ -1,9 +1,11 @@
 // TypeScript Version: 2.8
 import * as queries from './queries'
+import * as queryHelpers from './query-helpers'
 
-export {queries}
+export {queries, queryHelpers}
 
 export * from './queries'
+export * from './query-helpers'
 export * from './wait'
 export * from './wait-for-element'
 export * from './matches'
