+++ typings/query-helpers.d.ts
@@ -1,0 +1,28 @@
+import {Matcher, MatcherOptions} from './matches'
+
+export interface SelectorMatcherOptions extends MatcherOptions {
+  selector?: string
+}
+
+export type QueryByAttribute = (
+  container: HTMLElement,
+  id: Matcher,
+  options?: MatcherOptions,
+) => HTMLElement | null
+
+export type AllByAttribute = (
+  container: HTMLElement,
+  id: Matcher,
+  options?: MatcherOptions,
+) => HTMLElement[]
+
+export const queryByAttribute: QueryByAttribute
+export const queryAllByAttribute: AllByAttribute
+export const firstResultOrNull: (
+  fn: AllByAttribute,
+  container?: HTMLElement,
+  id?: Matcher,
+  options?: MatcherOptions,
+) => HTMLElement | null
+export const debugDOM: (htmlElement: HTMLElement) => string
+export const getElementError: (message: string, container: HTMLElement) => Error
